Created Custom Editor for VRControllerEvents Inspector called VRControllerEventsEditor Within this Editor I can easily select which variables I want to use in the scene for a particular controller. If I click set the bool useGravitationalPulsar to true then it will display all the variables and automatically activate that part of the code without errors.
I've also managed to fix the bug whereby the camera rig was deciding to fall over in game and end up upside down by removing the VR Tracking object script from the Camerarig gameobject.

I've also added in the sound variables to the VRControllerEvents script in preparation for when I create the sound effects
Likewise I've also added in automatic component require when attaching the VRControllerEvents script to any controller it will automatically add a Linerenderer & AudioSource to the gameobject.

I have been creating some grappler scripts tomorrow I will hopefully test them out.

I also need to set up the the movement script so that depending on the transform rotation of the headset on the camera will depend on which direction the player moves forwards. As well as sorting out the jump mechanics to the user so they must be grounded in order to be able to jump.

I also need to work on completing the 2nd iteration of the Gravitiational Pulsar.